scripts/requirements.txt
pytz==2023.3.post1
pyyaml==6.0.1